requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Flask-Limiter>=3.5.0
Flask-Caching>=2.1.0
Werkzeug==2.3.7
openpyxl==3.1.2
google-generativeai>=0.3.0
waitress==3.0.0
python-dotenv==1.0.0

# SQL Server Desteği
pyodbc>=5.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

# PDF Raporlama
reportlab>=4.0.0

# Güvenlik
Flask-Talisman>=1.1.0
python-magic>=0.4.27

# API Documentation
flask-restx>=1.3.0

# Webhook
requests>=2.31.0

# Veri Simülasyonu
Faker==24.0.0

# Görev Zamanlama (Scheduler)
APScheduler>=3.10.4
